feat(): eo-sidebar-menu
Closes NEXT_BUILDER-3794
