+++ bricks/basic/src/sidebar/sidebar-menu/index.tsx
@@ -1,9 +1,9 @@
-import React, { useCallback, useEffect, useRef } from "react";
+import React from "react";
 import { createDecorators } from "@next-core/element";
 import { ReactNextElement } from "@next-core/react-element";
 import "@next-core/theme";
 import styleText from "./styles.shadow.css";
-import { isSidebarMenuItem } from "../utils.js";
+import { useUpdateMenuCollapsedState } from "../utils.js";
 
 const { defineElement, property } = createDecorators();
 
@@ -33,31 +33,7 @@ class EoSidebarMenu extends ReactNextElement implements EoSidebarMenuProps {
 export function EoSidebarMenuComponent(props: EoSidebarMenuProps) {
   const { menuCollapsed } = props;
 
-  const slotRef = useRef<HTMLSlotElement>(null);
-
-  const updateChildrenMenuCollapsedState = useCallback(() => {
-    slotRef.current?.assignedElements().forEach((ele) => {
-      if (isSidebarMenuItem(ele)) {
-        (ele as any).menuCollapsed = menuCollapsed;
-      }
-    });
-  }, [menuCollapsed]);
-
-  useEffect(() => {
-    updateChildrenMenuCollapsedState();
-  }, [menuCollapsed, updateChildrenMenuCollapsedState]);
-
-  useEffect(() => {
-    const slotElem = slotRef.current;
-    const handleSlotchange = () => {
-      updateChildrenMenuCollapsedState();
-    };
-    slotElem?.addEventListener("slotchange", handleSlotchange);
-
-    return () => {
-      slotElem?.removeEventListener("slotchange", handleSlotchange);
-    };
-  }, [updateChildrenMenuCollapsedState]);
+  const [slotRef] = useUpdateMenuCollapsedState(menuCollapsed);
 
   return (
     <div className="sidebar-menu">
